Lb some final deets

+++ src/components/DeleteDirectoryButton.js
@@ -1,20 +1,22 @@
 import DeleteDirectory from '../hooks/DeleteDirectory';
 import CheckUser from '../hooks/CheckUser';
 
-const DeleteDirectoryButton = ({directory}) => {
+const DeleteDirectoryButton = ({ directory }) => {
 
     const deleteDirectory = async () => {
-        try {
-            await DeleteDirectory(directory, localStorage.getItem('userToken'));
-            alert('Successfully deleted category!');
-            window.location.href = '#/';
-            window.location.reload();
-        }
-        catch (error) {
-            alert('Failed to delete category: ', error);
+        if (window.confirm("Are you sure you want to delete this category? This action cannot be undone.")) {
+            try {
+                await DeleteDirectory(directory, localStorage.getItem('userToken'));
+                alert('Successfully deleted category!');
+                window.location.href = '#/';
+                window.location.reload();
+            }
+            catch (error) {
+                alert('Failed to delete category: ', error);
+            }
         }
     };
-    
+
     if (CheckUser()) {
         return <button className="rightButton" onClick={deleteDirectory}>Delete Category</button>;
     }
